requirements.txt
django>=5.2.1
django-crispy-forms>=2.4
crispy-tailwind>=1.0.3
gunicorn>=21.2.0
python-dotenv>=1.0.0
whitenoise>=6.5.0
django-allauth>=0.54.0
pillow>=10.0.0
django-tailwind>=3.6.0
reportlab>=4.0.0